requirements.dev.txt
flake8==3.8.1
mypy==0.770
black==19.10b0
pytest-django==3.9.0
pytest-xdist==1.32.0
pytest-factoryboy==2.0.3
pytest-cov==2.10.0
